Move DivestmentList component to InstitutionalDivestments in sections folder
